refactor: 도메인 명 변경
